# src/docs/swagger.yaml
definitions:
  api.AuthResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
      user_id:
        type: string
    type: object
...
  api.SignUpRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  api.UserResponse:
    properties:
      login:
        type: string
      user_id:
        type: string
    type: object
  domain.JwtRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  domain.JwtResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      type:
        description: Обычно "Bearer"
        type: string
    type: object
...
  domain.RefreshJwtRequest:
    properties:
      refreshToken:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/refresh-access:
    post:
      consumes:
      - application/json
      description: Принимает refreshToken и возвращает новый accessToken БЕЗ изменения
        сессии рефреша. Доступен без авторизации.
      parameters:
      - description: Действующий Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новый access токен
          schema:
            $ref: '#/definitions/domain.JwtResponse'
        "400":
          description: Некорректный формат JSON
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "401":
          description: Невалидный или просроченный токен
          schema:
            $ref: '#/definitions/api.AuthResponse'
      summary: Обновление Access токена
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Принимает refreshToken, инвалидирует его и возвращает полностью
        новую пару Access и Refresh токенов.
      parameters:
      - description: Действующий Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новая пара токенов
          schema:
            $ref: '#/definitions/domain.JwtResponse'
        "400":
          description: Некорректный формат JSON
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "401":
          description: Невалидный или просроченный токен
          schema:
            $ref: '#/definitions/api.AuthResponse'
      summary: Полное обновление пары токенов (Ротация)
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные пользователя (логин/пароль) и возвращает
        пару JWT токенов.
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.JwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход, возвращает токены
          schema:
            $ref: '#/definitions/domain.JwtResponse'
        "400":
          description: Некорректный формат JSON
          schema:
            $ref: '#/definitions/domain.JwtResponse'
        "401":
          description: Неверный логин или пароль
          schema:
            $ref: '#/definitions/domain.JwtResponse'
      summary: Авторизация пользователя (Вход)
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Создает новый аккаунт и сразу возвращает JWT-токены для авторизации.
      parameters:
      - description: Данные для регистрации (логин и пароль)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "400":
          description: Некорректное тело запроса (невалидный JSON)
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "422":
          description: Ошибка бизнес-валидации (пользователь уже существует, слишком
            короткий пароль)
          schema:
            $ref: '#/definitions/api.AuthResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
...
  /user/me:
    get:
      description: Возвращает профиль пользователя, извлекая UUID из проверенного
        accessToken в заголовке Authorization.
      produces:
      - application/json
      responses:
        "200":
          description: Данные профиля авторизованного пользователя
          schema:
            $ref: '#/definitions/api.UserResponse'
        "401":
          description: Токен отсутствует, невалиден или просрочен
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "404":
          description: Пользователь с таким UUID не найден в БД
          schema:
            $ref: '#/definitions/api.AuthResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о текущем пользователе
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {accessToken}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"